Refactor labwhere methods into own module
We're about to add a post action for
labwhere, which isn't used in the reports
so lets pull it into its own module
