+++ tests/helpers/test_reports_helpers.py
@@ -23,7 +23,6 @@ from lighthouse.helpers.reports import (
     get_distinct_plate_barcodes,
     get_new_report_name_and_path,
     join_samples_declarations,
-    map_labware_to_location,
     report_query_window_start,
     unpad_coordinate,
 )
@@ -171,33 +170,6 @@ def test_query_by_ct_limit(app, freezer, samples_ct_values):
         ct_less_than_limit = samples.count_documents({FIELD_CH1_CQ: {"$lte": CT_VALUE_LIMIT}})
 
         assert ct_less_than_limit == 1  # 'MCM003'
-
-
-def test_map_labware_to_location_labwhere_error(app, freezer, labwhere_samples_error):
-    # mocks response from get_locations_from_labwhere() with labwhere_samples_error
-
-    raised_exception = False
-
-    try:
-        with app.app_context():
-            map_labware_to_location(["test"])
-    except ReportCreationError:
-        raised_exception = True
-
-    assert raised_exception
-
-
-def test_map_labware_to_location_dataframe_content(app, freezer, labwhere_samples_multiple):
-    # mocks response from get_locations_from_labwhere() with labwhere_samples_multiple
-    with app.app_context():
-        result = map_labware_to_location(
-            ["123", "456", "789"]
-        )  # '789' returns no location, in the mock
-
-    expected_columns = [FIELD_PLATE_BARCODE, "location_barcode"]
-    expected_data = [["123", "4567"], ["456", "1234"], ["789", ""]]
-    assert result.columns.to_list() == expected_columns
-    assert np.array_equal(result.to_numpy(), expected_data)
 
 
 def test_add_cherrypicked_column(app, freezer):
